learning/Google ADK/requirements.txt
google-adk
litellm
yfinance
psutil
google-generativeai
python-dotenv